veille_aggregator_mvp/veille_aggregator_mvp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
feedparser==6.0.11
trafilatura==1.12.2
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
httpx==0.28.1
pytest==8.3.5
